Prevent exception for missing scheduler notes
It seems like it's only possible for notes to be missing if the user
deleted the section in a separate scheduler instance open concurrently.
We don't really support having multiple schedulers open for the same user,
but should handle the case gracefully by pretending the section has no notes.
The client-side code can already handle the notes being null.